Modified the output TSB

+++ src/core/task.ts
@@ -1,6 +1,5 @@
 import {CompileModuleData, Config, CopyData, RemoveData} from "./config";
-import {set_full_value, set_status, set_step_value, write_status_message} from "./output";
-import {Color} from "./console";
+import {set_full_value, set_status, set_step_value, write_error, write_status_message, write_warning} from "./output";
 import * as fs from "fs";
 import * as path from "path";
 import {BUILD_OPTIONS, CWD, ENGINE_DIR} from "./global";
@@ -9,7 +8,7 @@ import {EmitOutput, OutputFile, SourceFile} from "ts-morph";
 import {compile_module} from "./transpiler";
 import {Plugin, PluginHandler, PluginResultInformation} from "../plugin/plugin";
 
-export function compile_module_task(config: Config, information: CompileModuleData, index: number): void {
+export function compile_module_task(config: Config, information: CompileModuleData): void {
     if (!!config.modules[information.moduleName]) {
         const name: string = information.moduleName;
         const sources: string[] = config.modules[information.moduleName];
@@ -20,25 +19,28 @@ export function compile_module_task(config: Config, information: CompileModuleDa
             const component: Plugin | null = PluginHandler.instantiate(plugin.name, plugin.parameters);
 
             if (!component) {
-                write_status_message(index, `ERROR: Cannot find the plugin '${plugin.name}' found ${PluginHandler.names()}`)
-                set_status(index, "FAIL");
+                write_error(`ERROR: Cannot find the plugin '${plugin.name}' found ${PluginHandler.names()}`)
+                set_status("FAIL");
                 return;
             }
             plugins.push(component);
         }
 
-        const result: SourceFile = compile_module(name, sources, loaders, plugins, index);
+        const result: SourceFile | null = compile_module(name, sources, loaders, plugins);
+        if (result == null) {
+            return;
+        }
 
-        set_full_value(index, 84);
+        set_full_value(0.84);
         if (BUILD_OPTIONS.produceTS) {
-            write_status_message(index, "Create Typescript file");
+            write_status_message("Create Typescript file");
             fs.writeFileSync(result.getFilePath(), result.print());
         } else {
             fs.unlinkSync(path.join(CWD, "out", name + ".ts"));
         }
 
-        write_status_message(index, "Write output");
-        set_full_value(index, 96);
+        write_status_message("Write output");
+        set_full_value(0.96);
         let output: EmitOutput = result.getEmitOutput();
 
         let data: OutputFile = output.getOutputFiles().filter(value => value.getFilePath() == result.getFilePath().replace(path.extname(result.getFilePath()), ".js"))[0];
@@ -62,30 +64,30 @@ export function compile_module_task(config: Config, information: CompileModuleDa
 
         fs.writeFileSync(data.getFilePath(), transformedData);
 
-        set_status(index, "OK");
+        set_status("OK");
         return;
     }
 
-    write_status_message(index, "ERROR: Cannot find module '" + information.moduleName + "'", Color.Red);
-    set_status(index, "FAIL");
+    write_error("ERROR: Cannot find module '" + information.moduleName + "'");
+    set_status("FAIL");
 }
 
-export function copy_task(information: CopyData, index: number): void {
+export function copy_task(information: CopyData): void {
     if (!fs.existsSync(path.join(CWD, information.from))) {
-        write_status_message(index, "ERROR: Cannot find item '" + information.from + "'", Color.Red);
-        set_status(index, "FAIL");
+        write_error("ERROR: Cannot find item '" + information.from + "'");
+        set_status("FAIL");
         return;
     }
 
     if (!fs.existsSync(path.join(CWD, information.to)) || !fs.statSync(path.join(CWD, information.to)).isDirectory()) {
-        write_status_message(index, "ERROR: Cannot find item '" + information.to + "'", Color.Red);
-        set_status(index, "FAIL");
+        write_error("ERROR: Cannot find item '" + information.to + "'");
+        set_status("FAIL");
         return;
     }
 
     if (fs.statSync(path.join(CWD, information.from)).isFile()) {
         fs.copyFileSync(path.join(CWD, information.from), path.join(CWD, information.to, path.basename(information.from)));
-        set_status(index, "OK");
+        set_status("OK");
         return;
     }
 
@@ -95,78 +97,78 @@ export function copy_task(information: CopyData, index: number): void {
     }
 
 
-    write_status_message(index, "Collecting directories");
+    write_status_message("Collecting directories");
     let dirs: string[] = list_dirs(path.join(CWD, information.from));
-    write_status_message(index, "Collecting files");
-    set_full_value(index, 25);
+    write_status_message("Collecting files");
+    set_full_value(0.25);
     let files: string[] = list_files(path.join(CWD, information.from));
 
-    set_full_value(index, 50);
+    set_full_value(0.50);
     dirs.forEach((dir: string, value: number): void => {
-        set_step_value(index, Math.round(value * 100 / files.length));
-        write_status_message(index, `Create '${dir.replace(path.join(CWD, information.from), "")}'`);
+        set_step_value(value / dirs.length);
+        write_status_message(`Create '${dir.replace(path.join(CWD, information.from), "")}'`);
         if (!fs.existsSync(dir.replace(path.join(CWD, information.from), path.join(information.to, path.basename(information.from))))) {
             fs.mkdirSync(dir.replace(path.join(CWD, information.from), path.join(information.to, path.basename(information.from))));
         }
     });
 
-    set_full_value(index, 75);
+    set_full_value(0.75);
     files.forEach((file, value) => {
-        set_step_value(index, Math.round(value * 100 / files.length));
-        write_status_message(index, `Copy '${file.replace(path.join(CWD, information.from), "")}'`);
+        set_step_value(value / files.length);
+        write_status_message(`Copy '${file.replace(path.join(CWD, information.from), "")}'`);
         if (!fs.existsSync(file.replace(path.join(CWD, information.from), information.to)) || information.overwrite) {
             fs.copyFileSync(file, file.replace(path.join(CWD, information.from), path.join(information.to, path.basename(information.from))));
         }
 
     });
 
-    set_status(index, "OK");
+    set_status("OK");
 }
 
-export function remove_task(information: RemoveData, index: number): void {
+export function remove_task(information: RemoveData): void {
     if (fs.existsSync(path.join(CWD, information.target))) {
         if (fs.statSync(path.join(CWD, information.target)).isFile()) {
             fs.unlinkSync(path.join(CWD, information.target));
-            set_status(index, "OK");
+            set_status("OK");
             return;
         }
 
         if (fs.statSync(path.join(CWD, information.target)).isDirectory()) {
             if (information.recursive) {
-                write_status_message(index, "Collecting folders")
+                write_status_message("Collecting folders")
                 let dirs: string[] = list_dirs(path.join(CWD, information.target), false);
 
-                set_full_value(index, 25);
-                write_status_message(index, "Collecting files")
+                set_full_value(0.25);
+                write_status_message("Collecting files")
                 let files: string[] = list_files(path.join(CWD, information.target));
 
-                set_full_value(index, 50);
+                set_full_value(0.50);
                 files.forEach((file, value) => {
-                    set_step_value(index, Math.round(value * 100 / files.length));
-                    write_status_message(index, `Removing '${file.replace(path.join(CWD, information.target), "")}'`);
+                    set_step_value(value / files.length);
+                    write_status_message(`Removing '${file.replace(path.join(CWD, information.target), "")}'`);
                     fs.unlinkSync(file);
                 });
 
-                set_full_value(index, 75);
+                set_full_value(0.75);
                 dirs.forEach((dir, value) => {
-                    set_step_value(index, Math.round(value * 100 / dirs.length));
-                    write_status_message(index, `Removing '${dir.replace(path.join(CWD, information.target), "")}'`);
+                    set_step_value(value / dirs.length);
+                    write_status_message(`Removing '${dir.replace(path.join(CWD, information.target), "")}'`);
                     fs.rmdirSync(dir);
                 });
 
-                set_status(index, "OK");
+                set_status("OK");
                 return;
             } else {
-                set_status(index, "FAIL");
-                write_status_message(index, "ERROR: Item to remove is a folder but recursive flag is not set", Color.Red);
+                set_status("FAIL");
+                write_error("ERROR: Item to remove is a folder but recursive flag is not set");
                 return;
             }
         }
 
-        set_status(index, "FAIL");
-        write_status_message(index, "ERROR: '" + information.target + "' is not a file or a folder", Color.Red);
+        set_status("FAIL");
+        write_error("ERROR: '" + information.target + "' is not a file or a folder");
         return;
     }
-    write_status_message(index, "WARNING: Could not find '" + information.target + "'", Color.Yellow);
-    set_status(index, "WARNING");
+    write_warning("WARNING: Could not find '" + information.target + "'");
+    set_status("WARNING");
 }
